Fix cocoapods dependencies logic
Previously, VirtusizeAuth.framework was included as a "copy" in the
Virtusize project.
That was a reason, why the `pod lib lint` worked, even though the
reference of VirtusizeCore was incorrect.

After we changed the type of dependency of VirtusizeAuth, we simply fix
the incorrect podspec structure.

There was two options:
1. Reference VirtusizeCore and VirtusizeAuth as pure code (not as projects
or libraries -> so we need to omit `import VirtusizeCore` usage)
2. Make all 3 project as independant pods

So far we go with option 2 and each project will be a separate pod.
In the following release, we should probably migrate to option 1.